requirement.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
